Fixed overflow in cost matrix size calculation

+++ RAPTor/raytracer/spatial_sub_division/bvh_builder.h
@@ -55,7 +55,7 @@ class bvh_builder
         const int                       _max_node_size;
         const int                       _max_down_phase_depth;
         int                             _depth;
-        int                             _rows;
+        long                            _rows;
         std::atomic<int>                _next_node;
 };
 }; /* namespace raptor_raytracer */
